requirements.txt
python-dotenv==1.0.0
requests==2.31.0
duckdb==0.8.1
duckdb-engine==0.7.3
SQLAlchemy==2.0.16
ipython-sql==0.5.0
ipykernel==6.23.2
matplotlib==3.7.1
pandas==2.0.2